+++ tests/testthat/test-enrich_sfr.R
@@ -2,7 +2,7 @@ test_that("enrich_sfr works", {
 
     # Given below PPM values as input borders (3 points left, 2 points right),
     # the correct output would be DP1, but the expected output is DP2 to
-    # maintain backwards compatibility.
+    # maintain backwards compatibility with MetaboDecon1D (v0.2.2).
     #
     # PPM: 4.7   3.4   2.1 | 0.8  -0.5   -1.8 | -3.1  -4.4
     # DP1: 7.0   6.0   5.0 | 4.0   3.0    2.0 |  1.0   0.0
@@ -19,7 +19,7 @@ test_that("enrich_sfr works", {
     spec <- named(ppm, dp, n, ppm_min, ppm_max, ppm_step, ppm_nstep, sf)
     class(spec) <- "ispec"
 
-    sfr <- enrich_sfr(spec, sfr = c(2.0, -3.0))
+    sfr <- enrich_sfr(sfr = c(2.0, -3.0), x = spec)
     expect_equal(sfr, list(
         left_ppm  = 2,
         right_ppm = -3,
@@ -28,4 +28,8 @@ test_that("enrich_sfr works", {
         left_sdp  = 0.00662637362637363,
         right_sdp = 0.00223076923076923
     ))
+
+    sfr2 <- enrich_sfr2(sfr = c(2.0, -3.0), cs = ppm)
+    expect_equal(sfr, sfr2)
+
 })